+++ cases/testcases/deprecations/own_ns/green.clj
@@ -1,0 +1,7 @@
+(ns testcases.deprecations.own-ns.green)
+
+(defn ^:deprecated foo [])
+
+;; https://github.com/jonase/eastwood/issues/402
+(defn bar []
+  (foo))