+++ cases/testcases/deprecations/own_ns/red.clj
@@ -1,0 +1,6 @@
+(ns testcases.deprecations.own-ns.red
+  (:require
+   [testcases.deprecations.own-ns.green :as green]))
+
+(defn foo []
+  (green/foo))